+++ src/lib/pages/contract-details/components/tables/TxsTable.tsx
@@ -3,7 +3,7 @@ import { usePaginator } from "lib/components/pagination/usePaginator";
 import { EmptyState, ErrorFetching } from "lib/components/state";
 import { TransactionsTable } from "lib/components/table";
 import { DEFAULT_TX_FILTERS } from "lib/data";
-import { useTxsByAddress } from "lib/services/txService";
+import { useTxsByAddress } from "lib/services/tx";
 import type { BechAddr32, Option } from "lib/types";
 
 interface TxsTableProps {