+++ src/lib/pages/tx-details/components/TxHeader.tsx
@@ -6,7 +6,7 @@ import { useMobile } from "lib/app-provider";
 import { CopyLink } from "lib/components/CopyLink";
 import { CustomIcon } from "lib/components/icon";
 import { useOpenTxTab } from "lib/hooks";
-import type { TxData } from "lib/services/txService";
+import type { TxData } from "lib/services/types";
 import { dateFromNow, formatUTC } from "lib/utils";
 
 interface TxHeaderProps extends FlexProps {